+++ src/Export/Base/XlsxRenderer.php
@@ -1,0 +1,52 @@
+<?php
+
+/*
+ * This file is part of the Kimai time-tracking app.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace App\Export\Base;
+
+use PhpOffice\PhpSpreadsheet\IOFactory;
+use PhpOffice\PhpSpreadsheet\Spreadsheet;
+
+class XlsxRenderer extends AbstractSpreadsheetRenderer
+{
+    public function getFileExtension(): string
+    {
+        return '.xlsx';
+    }
+
+    /**
+     * @return string
+     */
+    protected function getContentType(): string
+    {
+        return 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet';
+    }
+
+    /**
+     * @param Spreadsheet $spreadsheet
+     * @return string
+     * @throws \Exception
+     */
+    protected function saveSpreadsheet(Spreadsheet $spreadsheet): string
+    {
+        $filename = tempnam(sys_get_temp_dir(), 'kimai-export-xlsx');
+        if (false === $filename) {
+            throw new \Exception('Could not open temporary file');
+        }
+
+        $writer = IOFactory::createWriter($spreadsheet, 'Xlsx');
+        $writer->save($filename);
+
+        return $filename;
+    }
+
+    public function getId(): string
+    {
+        return 'xlsx';
+    }
+}
