+++ src/Export/Renderer/CsvRenderer.php
@@ -9,73 +9,18 @@
 
 namespace App\Export\Renderer;
 
+use App\Export\Base\CsvRenderer as BaseCsvRenderer;
 use App\Export\RendererInterface;
-use PhpOffice\PhpSpreadsheet\IOFactory;
-use PhpOffice\PhpSpreadsheet\Spreadsheet;
-use PhpOffice\PhpSpreadsheet\Worksheet\Worksheet;
 
-final class CsvRenderer extends AbstractSpreadsheetRenderer implements RendererInterface
+final class CsvRenderer extends BaseCsvRenderer implements RendererInterface
 {
-    /**
-     * @return string
-     */
-    public function getFileExtension(): string
-    {
-        return '.csv';
-    }
-
-    /**
-     * @return string
-     */
-    protected function getContentType(): string
-    {
-        return 'text/csv';
-    }
-
-    /**
-     * @param Spreadsheet $spreadsheet
-     * @return string
-     * @throws \Exception
-     */
-    protected function saveSpreadsheet(Spreadsheet $spreadsheet): string
-    {
-        $filename = tempnam(sys_get_temp_dir(), 'kimai-export-csv');
-        if (false === $filename) {
-            throw new \Exception('Could not open temporary file');
-        }
-
-        $writer = IOFactory::createWriter($spreadsheet, 'Csv');
-        $writer->save($filename);
-
-        return $filename;
-    }
-
-    /**
-     * @return string
-     */
-    public function getId(): string
-    {
-        return 'csv';
-    }
-
-    /**
-     * @return string
-     */
     public function getIcon(): string
     {
         return 'csv';
     }
 
-    /**
-     * @return string
-     */
     public function getTitle(): string
     {
         return 'csv';
-    }
-
-    protected function setDuration(Worksheet $sheet, $column, $row, $duration)
-    {
-        $sheet->setCellValueByColumnAndRow($column, $row, sprintf('=%s', $duration));
     }
 }
