+++ tests/Controller/TimesheetTeamControllerTest.php
@@ -37,7 +37,10 @@ class TimesheetTeamControllerTest extends ControllerBaseTest
 
         $this->assertPageActions($client, [
             'search search-toggle visible-xs-inline' => '#',
-            'download toolbar-action' => $this->createUrl('/team/timesheet/export'),
+            'toolbar-action exporter-csv' => $this->createUrl('/team/timesheet/export/csv'),
+            'toolbar-action exporter-print' => $this->createUrl('/team/timesheet/export/print'),
+            'toolbar-action exporter-pdf' => $this->createUrl('/team/timesheet/export/pdf'),
+            'toolbar-action exporter-xlsx' => $this->createUrl('/team/timesheet/export/xlsx'),
             'visibility' => '#',
             'create modal-ajax-form' => $this->createUrl('/team/timesheet/create'),
             'help' => 'https://www.kimai.org/documentation/timesheet.html'
